Created a website landing page and blog page
This is a static website created to display information about the project. It is a website which has a blog integrated in it. The website can be found in https://soffyjay.github.io/mbaolab-landing/  and a sample blog post in https://soffyjay.github.io/mbaolab-landing/blog.html. The blog post is based on the article that was submitted in the pull request #50 at https://github.com/Mboalab/Mboalab-Outreachy_December-to-March-2022-internship-round/pull/50. More functionalities would be added as the project proceeds.